Move signature to declaration package
